+++ internal/domain/stock.go
@@ -1,7 +1,7 @@
 package domain
 
 type Stock struct {
-	SkuID      uint32
-	TotalCount uint64
-	Reserved   uint64
+	SkuID      int
+	TotalCount int
+	Reserved   int
 }
